+++ packages/core-kit/src/domains/analysis/error.ts
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) 2025.
+ * Author Peter Placzek (tada5hi)
+ * For the full copyright and license information,
+ * view the LICENSE file that was distributed with this source code.
+ */
+
+export class AnalysisError extends Error {
+    static defaultNodeRequired() {
+        return new AnalysisError('At least one default node must be selected.');
+    }
+
+    static aggregatorNodeRequired() {
+        return new AnalysisError('An aggregator node node must be selected.');
+    }
+
+    static imageAssignmentRequired() {
+        return new AnalysisError('A master image must be assigned to the analysis.');
+    }
+
+    static configurationLocked() {
+        return new AnalysisError('The analysis configuration is locked.');
+    }
+
+    static buildInitialized() {
+        return new AnalysisError('The analysis build process has already been initialized.');
+    }
+}