+++ packages/server-core/src/config/module.ts
@@ -17,6 +17,7 @@ import {
 } from '../aggregators/analysis-manager';
 import {
     createRegistryComponent,
+    useAnalysisConfigurationComponent,
 } from '../components';
 import {
     configureAmqp,
@@ -65,6 +66,7 @@ export function createConfig() : Config {
 
     const components : Component[] = [
         createRegistryComponent(),
+        useAnalysisConfigurationComponent(),
     ];
 
     return {