+++ packages/server-core/src/database/domains/analysis/subscriber.ts
@@ -6,7 +6,7 @@
  */
 
 import type {
-    EntitySubscriberInterface,
+    EntitySubscriberInterface, InsertEvent, UpdateEvent,
 } from 'typeorm';
 import { EventSubscriber } from 'typeorm';
 import {
@@ -15,6 +15,7 @@ import {
 import { BaseSubscriber } from '@privateaim/server-db-kit';
 import { EntityEventDestination } from '@privateaim/server-kit';
 import { DomainEventNamespace } from '@privateaim/kit';
+import { AnalysisConfigurationCommand, useAnalysisConfigurationComponent } from '../../../components';
 import { AnalysisEntity } from './entity';
 
 @EventSubscriber()
@@ -52,6 +53,32 @@ AnalysisEntity
                 return destinations;
             },
         });
+    }
+
+    async afterInsert(event: InsertEvent<AnalysisEntity>): Promise<any> {
+        await super.afterInsert(event);
+
+        if (event.entity.master_image_id) {
+            const analysisConfiguration = useAnalysisConfigurationComponent();
+            analysisConfiguration.trigger(
+                AnalysisConfigurationCommand.RECALC,
+                {
+                    analysisId: event.entity.id,
+                },
+            );
+        }
+    }
+
+    async afterUpdate(event: UpdateEvent<AnalysisEntity>): Promise<any> {
+        await super.afterUpdate(event);
+
+        const analysisConfiguration = useAnalysisConfigurationComponent();
+        analysisConfiguration.trigger(
+            AnalysisConfigurationCommand.RECALC,
+            {
+                analysisId: event.entity.analysis_id,
+            },
+        );
     }
 
     listenTo() {