+++ packages/server-core/test/unit/http/analysis/crud.spec.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2024.
+ * Copyright (c) 2021-2025.
  * Author Peter Placzek (tada5hi)
  * For the full copyright and license information,
  * view the LICENSE file that was distributed with this source code.
@@ -13,8 +13,8 @@ import {
     createTestSuite,
     expectPropertiesEqualToSrc,
     removeDateProperties,
-} from '../../utils';
-import { TEST_DEFAULT_ANALYSIS } from '../../utils/domains';
+} from '../../../utils';
+import { TEST_DEFAULT_ANALYSIS } from '../../../utils/domains';
 
 describe('src/controllers/core/analysis', () => {
     const suite = createTestSuite();