+++ packages/server-kit/src/core/component/handler/module.ts
@@ -10,9 +10,15 @@ import { isComponentHandlerFn } from './check';
 import type { ComponentHandler, ComponentHandlerFn } from './types';
 
 export class ComponentHandlers {
+    protected initializing : boolean;
+
+    protected initialized : boolean;
+
     protected handlers: Record<string, ComponentHandlerFn | ComponentHandler>;
 
     constructor() {
+        this.initializing = false;
+        this.initialized = false;
         this.handlers = {};
     }
 
@@ -24,18 +30,27 @@ export class ComponentHandlers {
         delete this.handlers[key];
     }
 
-    async setup() : Promise<void> {
+    async initialize() : Promise<void> {
+        if (this.initializing || this.initialized) {
+            return;
+        }
+
+        this.initializing = true;
+
         const keys = Object.keys(this.handlers);
         for (let i = 0; i < keys.length; i++) {
             const handler = this.handlers[keys[i]];
 
             if (
                 !isComponentHandlerFn(handler) &&
-                handler.setup
+                handler.initialize
             ) {
-                await handler.setup();
+                await handler.initialize();
             }
         }
+
+        this.initialized = true;
+        this.initializing = false;
     }
 
     async execute(
@@ -43,6 +58,8 @@ export class ComponentHandlers {
         value: ObjectLiteral = {},
         metadata: ObjectLiteral = {},
     ) : Promise<void> {
+        await this.initialize();
+
         const handler = this.handlers[key];
         if (!handler) {
             return;