+++ packages/core-kit/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [0.8.20](https://github.com/PrivateAIM/hub/compare/v0.8.19...v0.8.20) (2025-10-29)
+
+
+### Features
+
+* analysis aggregated configuration columns  ([#1267](https://github.com/PrivateAIM/hub/issues/1267)) ([e60c460](https://github.com/PrivateAIM/hub/commit/e60c460c1f701f8b73450e7c618d00de27f8462a))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @privateaim/kit bumped from ^0.8.19 to ^0.8.20
+    * @privateaim/telemetry-kit bumped from ^0.8.19 to ^0.8.20
+  * peerDependencies
+    * @privateaim/kit bumped from ^0.8.19 to ^0.8.20
+    * @privateaim/telemetry-kit bumped from ^0.8.19 to ^0.8.20
+
 ## [0.8.19](https://github.com/PrivateAIM/hub/compare/v0.8.18...v0.8.19) (2025-09-24)